Move new version to `v2` package.
