+++ scala-new/src/main/scala/com/github/fehu/opentracing/v2/internal/State.scala
@@ -1,4 +1,4 @@
-package com.github.fehu.opentracing.internal
+package com.github.fehu.opentracing.v2.internal
 
 import scala.jdk.CollectionConverters._
 
@@ -9,7 +9,7 @@ import cats.syntax.functor._
 import cats.syntax.traverse._
 import io.opentracing.{ Span, SpanContext, Tracer }
 
-import com.github.fehu.opentracing.Traced
+import com.github.fehu.opentracing.v2.Traced
 
 private[opentracing] case class State[F[_]](
   private[opentracing] val tracer: Tracer,
